Add files via upload
This is my second version of this project. 
In my first version, I calculate the inverse matrix in the first function and assign it to a global variable. Then retrieve it in second function. 
In this version, I only set the matrix in first function, don't do the calculate. I calculate the inverse matrix in second function